+++ test/browser/features/support/build-packages.rb
@@ -25,12 +25,14 @@ PACKAGE_NAMES = [
   "@bugsnag/core-performance",
   "@bugsnag/browser-performance",
   "@bugsnag/react-router-performance",
+  "@bugsnag/vue-router-performance",
 ]
 
 PACKAGE_DIRECTORIES = [
   "#{ROOT}/packages/core",
   "#{ROOT}/packages/platforms/browser",
   "#{ROOT}/packages/react-router",
+  "#{ROOT}/packages/vue-router",
 ]
 
 def run(command)